fix: 플레이어 오브젝트 키컨트롤 로직 분리
- 키보드 입력값에 따라 플레이어 오브젝트의 포지션, 방향 및 애니메이션 모션을 결정하는 로직을 별도의 파일로 분리
